+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendarScreen/calendarScreen/CalendarScreenViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 import kr.co.domain.usecase.calendar.GetUserCropsUseCase
 import kr.co.main.R
-import kr.co.main.mapper.calendar.CropModelMapper
+import kr.co.main.calendar.mapper.CropModelMapper
 import kr.co.main.calendar.model.CropModel
 import kr.co.main.calendar.ui.calendarScreen.calendarScreen.CalendarScreenViewModel.CalendarScreenState
 import kr.co.ui.base.BaseViewModel
