+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendarScreen/calendarScreen/scheduleTab/ScheduleTabViewModel.kt
@@ -9,9 +9,9 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.update
 import kr.co.domain.usecase.calendar.GetFarmWorksUseCase
 import kr.co.domain.usecase.calendar.GetHolidaysUseCase
-import kr.co.main.mapper.calendar.CropModelMapper
-import kr.co.main.mapper.calendar.FarmWorkModelMapper
-import kr.co.main.mapper.calendar.HolidayModelMapper
+import kr.co.main.calendar.mapper.CropModelMapper
+import kr.co.main.calendar.mapper.FarmWorkModelMapper
+import kr.co.main.calendar.mapper.HolidayModelMapper
 import kr.co.main.calendar.model.CropModel
 import kr.co.main.calendar.model.FarmWorkModel
 import kr.co.main.calendar.model.HolidayModel
